calculating weights by geometric distance

+++ DualContouring/octree.h
@@ -44,7 +44,7 @@ struct OctreeDrawInfo
 
     int index;
     int corners;
-    vm::vec4 biome;
+    vm::ivec4 biome;
     vm::vec4 biomeWeights;
     vm::vec3 position;
     vm::vec3 averageNormal;
